Move BackupRangePartitionedProgress* files to fdbserver/backupworker
